id deletion implemented

+++ src/main/java/com/globaltradecorp/staticIdTool/dao/AppUserRowMapper.java
@@ -13,7 +13,7 @@ public class AppUserRowMapper implements RowMapper<AppUser> {
     @Override
     public AppUser mapRow(ResultSet resultSet, int i) throws SQLException {
         return AppUser.builder()
-                .id(resultSet.getLong("id"))
+                .id(resultSet.getInt("id"))
                 .username(resultSet.getString("username"))
                 .email(resultSet.getString("email"))
                 .firstName(resultSet.getString("first_name"))
